requirements.txt
Babel==2.2.0
Django==1.8.3
MySQL-python==1.2.5
PySocks==1.5.6
SQLAlchemy==1.0.11
amqp==1.4.8
anyjson==0.3.3
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.22
celery==3.1.19
certifi==2015.11.20.1
django-celery==3.1.17
django-kombu==0.9.4
django-suit==0.2.16
eventlet==0.18.1
flower==0.8.4
futures==3.0.3
gevent==1.1rc3
greenlet==0.4.9
kombu==3.0.32
python-nmap==0.5.0-1
pytz==2015.7
requests==2.9.1
singledispatch==3.4.0.3
six==1.10.0
termcolor==1.1.0
tornado==4.2
wsgiref==0.1.2